+++ fsutil/fs_test.go
@@ -16,7 +16,7 @@ import (
 
 	"pkg.re/essentialkaos/ek.v12/system"
 
-	check "pkg.re/check.v1"
+	check "pkg.re/essentialkaos/check.v1"
 )
 
 // ////////////////////////////////////////////////////////////////////////////////// //