+++ knf/validators/fs/validators_test.go
@@ -16,7 +16,7 @@ import (
 
 	"pkg.re/essentialkaos/ek.v12/knf"
 
-	. "pkg.re/check.v1"
+	. "pkg.re/essentialkaos/check.v1"
 )
 
 // ////////////////////////////////////////////////////////////////////////////////// //