+++ usage/man/man_test.go
@@ -15,7 +15,7 @@ import (
 	"pkg.re/essentialkaos/ek.v12/timeutil"
 	"pkg.re/essentialkaos/ek.v12/usage"
 
-	. "pkg.re/check.v1"
+	. "pkg.re/essentialkaos/check.v1"
 )
 
 // ////////////////////////////////////////////////////////////////////////////////// //