+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * `go-check` package replaced by [our fork](https://kaos.sh/check)
 * `[cron]` Removed useless example
+* `[knf/validators/fs]` Fixed bug with handling pattern matching error
 
 ### 12.20.2