+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Changelog
 
+### 12.20.3
+
+* `go-check` package replaced by [our fork](https://kaos.sh/check)
+* `[cron]` Removed useless example
+* `[knf/validators/fs]` Fixed bug with handling pattern matching error
+* `[path]` Fixed usage examples
+
 ### 12.20.2
 
 * `[terminal]` Usage examples improvements